CDC surveillance network and network catchment area

What the CDC surveillance network covers

cdc_catchments

Format

A data frame with 17 rows and 2 variables:

name

character Network name

area

character Area

Source

Courtesy of Bob Rudis's cdccovidview package

Details

Table: Data summary

Name	cdc_catchments

Number of rows	17

Number of columns	2

_______________________

Column type frequency:

character	2

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

name	0	1	3	9	0	3	0

area	0	1	4	14	0	15	0

The Coronavirus Disease 2019 (COVID-19)-Associated Hospitalization Surveillance Network (COVID-NET) conducts population-based surveillance for laboratory-confirmed COVID-19-associated hospitalizations in children (persons younger than 18 years) and adults. The current network covers nearly 100 counties in the 10 Emerging Infections Program (EIP) states (CA, CO, CT, GA, MD, MN, NM, NY, OR, and TN) and four additional states through the Influenza Hospitalization Surveillance Project (IA, MI, OH, and UT). The network represents approximately 10% of US population (~32 million people). Cases are identified by reviewing hospital, laboratory, and admission databases and infection control logs for patients hospitalized with a documented positive SARS-CoV-2 test. Data gathered are used to estimate age-specific hospitalization rates on a weekly basis and describe characteristics of persons hospitalized with COVID-19. Laboratory confirmation is dependent on clinician-ordered SARS-CoV-2 testing. Therefore, the unadjusted rates provided are likely to be underestimated as COVID-19-associated hospitalizations can be missed due to test availability and provider or facility testing practices. COVID-NET hospitalization data are preliminary and subject to change as more data become available. All incidence rates are unadjusted. Please use the following citation when referencing these data: “COVID-NET: COVID-19-Associated Hospitalization Surveillance Network, Centers for Disease Control and Prevention. WEBSITE. Accessed on DATE”.

name	area

COVID-NET	Entire Network

EIP	California

EIP	Colorado

EIP	Connecticut

EIP	Entire Network

EIP	Georgia

EIP	Maryland

EIP	Minnesota

EIP	New Mexico

EIP	New York

EIP	Oregon

EIP	Tennessee

IHSP	Entire Network

IHSP	Iowa

IHSP	Michigan

IHSP	Ohio

IHSP	Utah

References

https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html

Author

Kieran Healy