CDC Surveillance Network Death Counts by Age

Provisional Death Counts for Coronavirus Disease (COVID-19)

cdc_deaths_by_age

Format

A data frame with 12 rows and 10 variables:

data_as_of

date When the data were most recently recorded

age_group

character Age range

start_week

date Start week

end_week

date End week

covid_deaths

integer COLUMN_DESCRIPTION

total_deaths

integer COLUMN_DESCRIPTION

percent_expected_deaths

double COLUMN_DESCRIPTION

pneumonia_deaths

integer COLUMN_DESCRIPTION

pneumonia_and_covid_deaths

integer COLUMN_DESCRIPTION

all_influenza_deaths_j09_j11

integer COLUMN_DESCRIPTION

Source

Courtesy of Bob Rudis's cdccovidview package

Details

Table: Data summary

Name	cdc_deaths_by_age

Number of rows	12

Number of columns	10

_______________________

Column type frequency:

character	1

Date	3

numeric	6

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

age_group	0	1	5	10	0	12	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

data_as_of	0	1	2020-04-30	2020-04-30	2020-04-30	1

start_week	0	1	2020-02-01	2020-02-01	2020-02-01	1

end_week	0	1	2020-04-25	2020-04-25	2020-04-25	1

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

covid_deaths	0	1	5753.50	9877.31	2.00	30.25	1211.50	7918.25	34521.00	▇▃▁▁▁

total_deaths	0	1	118897.67	202377.07	712.00	5675.25	28460.00	149341.50	713386.00	▇▂▁▁▁

percent_expected_deaths	0	1	0.97	0.00	0.97	0.97	0.97	0.97	0.97	▁▁▇▁▁

pneumonia_deaths	0	1	10454.17	18036.25	33.00	109.00	1799.50	14114.25	62725.00	▇▃▁▁▁

pneumonia_and_covid_deaths	0	1	2550.17	4387.93	0.00	12.50	491.50	3515.75	15301.00	▇▃▁▁▁

all_influenza_deaths_j09_j11	0	1	970.17	1618.90	11.00	40.75	358.50	1222.75	5821.00	▇▃▁▁▁

The U.S. Centers for Disease Control provides weekly summary and interpretation of key indicators that have been adapted to track the COVID-19 pandemic in the United States. Data is retrieved using the cdccovidview package from both COVIDView (https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html) and COVID-NET (https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html). Please see the indicated reference for all the caveats and precise meanings for each field.

References

https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true

Author

Kieran Healy