CDC Provisional death counts by week

Provisional Death Counts for Coronavirus Disease (COVID-19)

cdc_deaths_by_week

Format

A data frame with 13 rows and 10 variables:

data_as_of

date When the data were most recently recorded

start_week

date Start week

end_week

double End week

covid_deaths

integer COVID deaths

total_deaths

integer Total deaths

percent_expected_deaths

double COLUMN_DESCRIPTION

pneumonia_deaths

integer COLUMN_DESCRIPTION

pneumonia_and_covid_deaths

integer COLUMN_DESCRIPTION

all_influenza_deaths_j09_j11

integer COLUMN_DESCRIPTION

pneumonia_influenza_and_covid_19_deaths

integer COLUMN_DESCRIPTION

Source

Courtesy of Bob Rudis's cdccovidview package

Details

Table: Data summary

Name	cdc_deaths_by_week

Number of rows	13

Number of columns	10

_______________________

Column type frequency:

Date	3

numeric	7

________________________

Group variables	None

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

data_as_of	0	1	2020-04-30	2020-04-30	2020-04-30	1

start_week	0	1	2020-02-01	2020-04-25	2020-03-14	13

end_week	0	1	2020-02-01	2020-04-25	2020-03-14	13

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

covid_deaths	0	1	2655.46	4194.37	0.00	0.00	49.00	2659.00	11864.00	▇▁▁▂▁

total_deaths	0	1	54875.85	9864.46	24387.00	53940.00	56831.00	57299.00	65676.00	▁▁▁▇▂

percent_expected_deaths	0	1	0.97	0.17	0.45	0.97	0.97	0.99	1.19	▁▁▁▇▂

pneumonia_deaths	0	1	4825.00	2217.19	2219.00	3671.00	3692.00	5598.00	9580.00	▇▃▁▁▂

pneumonia_and_covid_deaths	0	1	1177.00	1863.76	0.00	0.00	25.00	1220.00	5281.00	▇▁▁▂▁

all_influenza_deaths_j09_j11	0	1	447.77	156.19	58.00	427.00	494.00	536.00	619.00	▁▁▁▇▇

pneumonia_influenza_and_covid_19_deaths	0	1	6690.23	4292.62	3553.00	4165.00	4275.00	7397.00	16272.00	▇▁▁▂▁

The U.S. Centers for Disease Control provides weekly summary and interpretation of key indicators that have been adapted to track the COVID-19 pandemic in the United States. Data is retrieved using the cdccovidview package from both COVIDView (https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html) and COVID-NET (https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html). Please see the indicated reference for all the caveats and precise meanings for each field.

References

https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true

Author

Kieran Healy