COVID-19 data for the USA, current as of Sunday, October 24, 2021

A dataset containing US state-level data on COVID-19

covus

Format

A tibble with 664,960 rows and 7 columns

date

Date in YYYY-MM-DD format (date)

state

Two letter State abbreviation (character)

fips

State FIPS code (character)

data_quality_grade

character Data quality as assessed by COVID Tracking Project staff

measure

Outcome measure for this date

count

Count of measure

measure_label

character Outcome measure, suitable for use as a plot label

Source

The COVID-19 Tracking Project https://covidtracking.com

Details

Table: Data summary

Name	covus

Number of rows	664960

Number of columns	7

_______________________

Column type frequency:

character	4

Date	1

logical	1

numeric	1

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1	2	2	0	56	0

fips	0	1	2	2	0	56	0

measure	0	1	5	30	0	31	0

measure_label	0	1	6	54	0	32	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

date	0	1	2020-01-13	2021-03-07	2020-09-03	420

Variable type: logical

skim_variable	n_missing	complete_rate	mean	count

data_quality_grade	664960	0	NaN	:

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

count	434334	0.35	387381.6	1638404	0	497	7770	134185.8	49646014	▇▁▁▁▁

The measures tracked by the COVID tracking project are as follows:

measure	measure_label

positive	Positive Tests

probable_cases	Probable Cases

negative	Negative Tests

pending	Pending Tests

hospitalized_currently	Currently Hospitalized

hospitalized_cumulative	Cumulative Hospitalized

in_icu_currently	Currently in ICU

in_icu_cumulative	Cumulative in ICU

on_ventilator_currently	Currently on Ventilator

on_ventilator_cumulative	Cumulative on Ventilator

recovered	Recovered

death	Deaths

hospitalized_discharged	Total Discharged from Hospital

total_tests_viral	Total number of PCR tests performed

positive_tests_viral	Total number of positive PCR tests

negative_tests_viral	Total number of negative PCR tests

positive_cases_viral	Total number of positive cases measured with PCR tests

death_confirmed	Deaths Confirmed

death_probable	Deaths Probable

total_test_encounters_viral	Total Test Encounters (PCR)

total_tests_people_viral	Total PCR Tests (People)

total_tests_antibody	Total Antibody Tests

positive_tests_antibody	Positive Antibody Tests

negative_tests_antibody	Total number of negative antibody tests

negative_tests_antibody	Negative Antibody Tests

total_tests_people_antibody	Total Antibody Tests (People)

positive_tests_people_antibody	Positive Antibody Tests (People)

negative_tests_people_antibody	Negative Antibody Tests (People)

total_tests_people_antigen	Total Antigen Tests (People)

positive_tests_people_antigen	Positive Antigen Tests (People)

total_tests_antigen	Total Antigen Tests

positive_tests_antigen	Positive Antigen Tests

Not all measures are reported by all states. The positive, negative, death, death_confirmed, probable_cases and death_probable measures are cumulative counts. death_confirmed is the total number deaths of individuals with COVID-19 infection confirmed by a laboratory test. In states where the information is available, it tracks only those laboratory-confirmed deaths where COVID also contributed to the death according to the death certificate. death_probable is the total number of deaths where COVID was listed as a cause of death and there is not a laboratory test confirming COVID-19 infection.

For further information on the COVID Tracking Project's measures, see https://covidtracking.com/about-data/data-definitions