COVID-19 case and death counts for the USA by race and state current as of Sunday, January 22, 2023

The COVID Racial Data Tracker advocates for, collects, publishes, and analyzes racial data on the pandemic across the United States. It’s a collaboration between the COVID Tracking Project and the Boston University Center for Antiracist Research.

covus_race

Format

A tibble with 47,880 rows and 7 columns

date

date Data reported as of this date

state

character State

group

character Racial group

cases

integer Total cases, count

deaths

integer Total deaths, count

hosp

integer Total hospitalizations, count

Source

https://covidtracking.com/race

Details

Table: Data summary

Name	covus_race

Number of rows	47880

Number of columns	7

_______________________

Column type frequency:

Date	1

character	2

numeric	4

________________________

Group variables	None

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

date	0	1	2020-04-12	2021-03-07	2020-09-23	95

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1	2	2	0	56	0

group	0	1	5	11	0	9	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

cases	15684	0.67	30240.68	103176.64	0	568	3661	21026	2619476	▇▁▁▁▁

deaths	17686	0.63	708.93	1836.84	-1	12	68	440	24402	▇▁▁▁▁

hosp	37253	0.22	2077.78	4654.37	0	67	345	1716	41099	▇▁▁▁▁

tests	43549	0.09	349773.42	1269936.08	0	6298	36108	199214	18567612	▇▁▁▁▁

The group variable is coded as follows:

groups

White

Black

Latino

Asian

AI/AN

NH/PI

Multiracial

Other

Unknown

AI/AN is American Indian/Alaska Native. NH/PI is Native Hawaiian/Pacific Islander. State-level counts should be handled with care, given the widely varying population distribution of people of different racial backgrounds by state.

Author

Kieran Healy