Google Mobility Data

Data from Google's Community Mobility Reports on relative changes in movement trends by location type.

google_mobility

Format

A data frame with 4,093,596 rows and 9 variables:

country_region_code

character Country Code

country_region

character Country or Region name

sub_region_1

character Subregion (e.g. US state) name

sub_region_2

character Subregion (e.g. US county) name

iso3166_2

character ISO 3166-2 Country/Region code

census_fips_code

character US Census FIPS code

date

double Date in yyyy-mm-dd format

type

character Type of location. Values are retail, grocery (and pharmacy), parts, transit (hubs/stations), workplaces, and residential

pct_diff

integer Percent change from baseline activity

Source

Google LLC "Google COVID-19 Community Mobility Reports." https://www.google.com/covid19/mobility/ Accessed: 2020-07-20

Details

Table: Data summary

Name	google_mobility

Number of rows	4093596

Number of columns	9

_______________________

Column type frequency:

character	7

Date	1

numeric	1

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

country_region_code	7044	1.00	2	2	0	134	0

country_region	0	1.00	4	22	0	135	0

sub_region_1	121500	0.97	3	74	0	1844	0

sub_region_2	1730148	0.58	5	27	0	1716	0

iso3166_2	2606028	0.36	4	6	0	1703	0

census_fips_code	1729242	0.58	5	5	0	2833	0

type	0	1.00	5	11	0	6	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

date	0	1	2020-02-15	2020-07-14	2020-04-29	151

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

pct_diff	1360864	0.67	-12.61	31.83	-100	-32	-9	5	625	▇▁▁▁▁

Location accuracy and the understanding of categorized places varies from region to region, so Google does not recommend using this data to compare changes between countries, or between regions with different characteristics (e.g. rural versus urban areas). Regions or categories are omitted if Google does not have have sufficient statistically significant levels of data for it. Changes for each day are compared to a baseline value for that day of the week. The baseline is the median value, for the corresponding day of the week, during the 5-week period Jan 3–Feb 6, 2020. What data is included in the calculation depends on user settings, connectivity, and whether it meets our privacy threshold. If the privacy threshold isn’t met (when somewhere isn’t busy enough to ensure anonymity) we don’t show a change for the day. As a result, you may encounter empty fields for certain places and dates. We calculate these insights based on data from users who have opted-in to Location History for their Google Account, so the data represents a sample of our users. As with all samples, this may or may not represent the exact behavior of a wider population.

Author

Kieran Healy