COVID-19 Case Surveillance Public Use Data

Deidentified patient-level data reported to U.S. states and autonomous reporting entities, including New York City and the District of Columbia (D.C.), as well as U.S. territories and states.

nchs_pud

Format

A tibble with 1,748,965 rows and 11 variables:

cdc_report_dt

date Initial case report date to CDC

pos_spec_dt

date Date of first positive specimen collection

onset_dt

date Symptom onset date, if symptomatic

current_status

character Case Status: Laboratory-confirmed case or Probable case

sex

character Sex: Male; Female; Unknown; Other

age_group

character Age Group: 0-9 Years; 10-19 Years; 20-39 Years; 40-49 Years; 50-59 Years; 60-69 Years; 70-79 Years; 80+ Years

race_ethnicity

character Race and ethnicity (combined): Hispanic/Latino; American Indian / Alaska Native, Non-Hispanic; Asian, Non-Hispanic; Black, Non-Hispanic; Native Hawaiian / Other Pacific Islander, Non-Hispanic; White, Non-Hispanic; Multiple/Other, Non-Hispanic

hosp_yn

character Hospitalization status

icu_yn

character ICU admission status

death_yn

character Death status

medcond_yn

character Presence of underlying comorbidity or disease

Source

https://data.cdc.gov/Case-Surveillance/COVID-19-Case-Surveillance-Public-Use-Data/vbim-akqf

Details

Table: Data summary

Name	nchs_pud

Number of rows	1748965

Number of columns	11

_______________________

Column type frequency:

character	8

Date	3

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

current_status	0	1	8	9	0	2	0

sex	15	1	4	7	0	4	0

age_group	38	1	6	8	0	10	0

race_ethnicity	9	1	7	28	0	8	0

hosp_yn	0	1	2	7	0	3	0

icu_yn	0	1	2	7	0	3	0

death_yn	0	1	2	7	0	3	0

medcond_yn	0	1	2	7	0	3	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

cdc_report_dt	0	1.00	2020-01-01	2020-06-11	2020-05-01	116

pos_spec_dt	1052553	0.40	2020-01-04	2020-06-24	2020-04-18	137

onset_dt	969687	0.45	2020-01-01	2020-06-26	2020-04-16	178

Each row is a deidentified patient. These deidentified data include demographic characteristics, exposure history, disease severity indicators and outcomes, clinical data, laboratory diagnostic test results, and comorbidities. All data elements can be found on the COVID-19 case report form located at www.cdc.gov/coronavirus/2019-ncov/downloads/pui-form.pdf. The COVID-19 case surveillance data are dynamic; case reports can be modified at any time by the reporting jurisdiction as new information becomes available (i.e., data are subject to change). Furthermore, reporting jurisdictions may report cases late. Version updates to the detailed and limited datasets will be available for request once a month. The datasets will include all cases with an initial report date of case to CDC at least 30 days prior to the creation of the previously updated datasets. This month lag will allow case reporting to be stabilized and ensure that time-dependent outcome data, including death, are accurately captured. Questions that have been left unanswered (blank) on the case report form are re-classified to an Unknown value, if applicable to the question. For example, in the question “Was the patient hospitalized?”, where the possible answer choices include “Yes”, “No”, or “Unknown”, the missing value is re-coded to the Unknown answer option if the respondent did not answer the question. The initial report date of the case to CDC is intended to be completed by the reporting jurisdiction when data are submitted. If blank, this variable is completed using the date the data file was first submitted to CDC. To prevent release of data that could be used to identify persons, data cells are suppressed for low frequency (<5) records and indirect identifiers (date of first positive specimen). Suppression includes uncommon combinations of demographic characteristics (sex, age group, race/ethnicity). Suppressed values are re-coded to the NA answer option.

Author

Kieran Healy