Provisional COVID-19 Death Counts by Sex, Age, and State

Deaths involving coronavirus disease (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.

nchs_sas

Format

A tibble with 85,374 rows and 15 variables:

data_as_of

date Date of data release

start_date

date First date of data period

end_date

date Last date of data period

group

character Unit of time observation: whether data in this row are measured By month, By total, or By year

year

integer Year of observation

month

integer Month of observation

state

character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, and New York City, separate from New York state.

sex

character Sex

age_group

character Age group

covid_19_deaths

integer Deaths involving COVID-19 (ICD-code U07.1)

total_deaths

integer Deaths from all causes of death

pneumonia_deaths

integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)

pneumonia_and_covid_19_deaths

integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)

influenza_deaths

integer Influenza Deaths (ICD-10 codes J09-J11)

pneumonia_influenza_or_covid_19_deaths

integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)

Source

National Center for Health Statistics https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku

Details

Table: Data summary

Name	nchs_sas

Number of rows	85374

Number of columns	15

_______________________

Column type frequency:

character	6

Date	1

numeric	8

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

start_date	0	1	10	10	0	27	0

end_date	0	1	10	10	0	27	0

group	0	1	7	8	0	3	0

state	0	1	4	20	0	54	0

sex	0	1	4	9	0	3	0

age_group	0	1	8	17	0	17	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

data_as_of	0	1	2022-03-16	2022-03-16	2022-03-16	1

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

year	2754	0.97	2020.70	0.69	2020	2020	2021	2021	2022	▇▁▇▁▂

month	11016	0.87	6.00	3.56	1	3	6	9	12	▇▃▃▃▅

covid_19_deaths	21441	0.75	409.04	6470.59	0	0	12	77	967044	▇▁▁▁▁

total_deaths	12424	0.85	2790.26	46372.88	0	42	154	681	7451652	▇▁▁▁▁

pneumonia_deaths	25425	0.70	382.21	5875.05	0	0	19	88	848894	▇▁▁▁▁

pneumonia_and_covid_19_deaths	21238	0.75	212.42	3378.14	0	0	0	38	501202	▇▁▁▁▁

influenza_deaths	14370	0.83	3.75	75.73	0	0	0	0	10618	▇▁▁▁▁

pneumonia_influenza_or_covid_19_deaths	24576	0.71	587.19	9066.19	0	0	28	132	1323785	▇▁▁▁▁

Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis, and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have been suppressed in accordance with NHCS confidentiality standards. As of September 2, 2020, this data file includes the following age groups in addition to the age groups that are routinely included: 0-17, 18-29, 30-49, and 50-64. The new age groups are consistent with categories used across CDC COVID-19 surveillance pages. When analyzing the file, the user should make sure to select only the desired age groups. Summing across all age categories provided will result in double counting deaths from certain age groups. Similarly, the state variable includes the United States as a whole, and New York City counted separately from the rest of New York State. The temporal unit of observation also varies, with totals given by year, by month, and overall. It is necessary to first filter the data by desired time unit, region, and age group to ensure there is no double-counting in subsequent calculations.

References

https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku

Author

Kieran Healy