NSSP National COVID-related ER Visits

National Syndromic Surveillance Program (NSSP): Emergency Department Visits and Percentage of Visits for COVID-19-Like Illness (CLI) or Influenza-like Illness (ILI)

nssp_covid_er_nat

Format

A data frame with 54 rows and 9 variables:

week

integer COLUMN_DESCRIPTION

num_fac

integer COLUMN_DESCRIPTION

total_ed_visits

character COLUMN_DESCRIPTION

visits

integer COLUMN_DESCRIPTION

pct_visits

double COLUMN_DESCRIPTION

visit_type

character COLUMN_DESCRIPTION

region

character COLUMN_DESCRIPTION

source

character COLUMN_DESCRIPTION

year

integer COLUMN_DESCRIPTION

Source

Courtesy of Bob Rudis's cdccovidview package

Details

Table: Data summary

Name	nssp_covid_er_nat

Number of rows	54

Number of columns	9

_______________________

Column type frequency:

character	4

numeric	5

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

total_ed_visits	0	1	7	7	0	27	0

visit_type	0	1	3	3	0	2	0

region	0	1	8	8	0	1	0

source	0	1	21	21	0	1	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

week	0	1	26.04	19.81	1.00	7.25	14.00	45.75	52.00	▇▂▁▂▇

num_fac	0	1	3346.89	48.97	3249.00	3329.50	3352.00	3389.50	3406.00	▃▁▆▃▇

visits	0	1	41521.67	16344.25	17639.00	31216.00	39183.50	50532.00	86088.00	▅▇▃▂▁

pct_visits	0	1	0.02	0.01	0.01	0.01	0.02	0.02	0.05	▇▆▂▁▂

year	0	1	2019.52	0.50	2019.00	2019.00	2020.00	2020.00	2020.00	▇▁▁▁▇

The U.S. Centers for Disease Control provides weekly summary and interpretation of key indicators that have been adapted to track the COVID-19 pandemic in the United States. Data is retrieved using the cdccovidview package from both COVIDView (https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html) and COVID-NET (https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html).

References

https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/04102020/nssp-regions.html

Author

Kieran Healy