NYT Excess Mortality Estimates, current as of Monday, July 6, 2020

All-cause mortality is widely used by demographers and other researchers to understand the full impact of deadly events, including epidemics, wars and natural disasters. The totals in this data include deaths from Covid-19 as well as those from other causes, likely including people who could not be treated or did not seek treatment for other conditions.

nytexcess

Format

A tibble with 6,712 rows and 12 columns

country

character Country Name

placename

character Place Name

frequency

character Reporting period. Weekly or monthly, depending on how the data is recorded.

start_date

date The first date included in the period.

end_date

date The last date included in the period,

year

character Year of data. Note that this variable is of type character and not integer because several observations are notes to the effect that the year is an average of two years.

month

integer Numerical month.

week

integer Numerical week.

deaths

integer The total number of confirmed deaths recorded from any cause.

expected_deaths

integer The baseline number of expected deaths, calculated from a historical average. See details below.

excess_deaths

integer The number of deaths minus the expected deaths.

baseline

character The years used to calculate expected_deaths.

Source

The New York Times https://github.com/nytimes/covid-19-data/tree/master/excess-deaths.

Details

expected deaths for each area based on historical data for the same time of year. These expected deaths are the basis for our excess death calculations, which estimate how many more people have died this year than in an average year.

The number of years used in the historical averages changes depending on what data is available, whether it is reliable and underlying demographic changes. See Data Sources for the years used to calculate the baselines. The baselines do not adjust for changes in age or other demographics, and they do not account for changes in total population.

The number of expected deaths are not adjusted for how non-Covid-19 deaths may change during the outbreak, which will take some time to figure out. As countries impose control measures, deaths from causes like road accidents and homicides may decline. And people who die from Covid-19 cannot die later from other causes, which may reduce other causes of death. Both of these factors, if they play a role, would lead these baselines to understate, rather than overstate, the number of excess deaths.

References

For further details on these data see https://github.com/nytimes/covid-19-data/tree/master/excess-deaths

Author

Kieran Healy