CDC provisional death counts by state

CDC Surveillance Network provisional death counts

cdc_deaths_by_state

Format

A data frame with 53 rows and 10 variables:

data_as_of

date Date most recently updated

state

character State name

start_week

date Start week

end_week

double End week

covid_deaths

integer COVID Deaths

total_deaths

integer Total deaths

percent_expected_deaths

double COLUMN_DESCRIPTION

pneumonia_deaths

integer COLUMN_DESCRIPTION

pneumonia_and_covid_deaths

integer COLUMN_DESCRIPTION

all_influenza_deaths_j09_j11

integer COLUMN_DESCRIPTION

Details

Table: Data summary

Name	cdc_deaths_by_state

Number of rows	53

Number of columns	10

_______________________

Column type frequency:

character	1

Date	3

numeric	6

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1	4	20	0	53	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

data_as_of	0	1	2020-04-30	2020-04-30	2020-04-30	1

start_week	0	1	2020-02-01	2020-02-01	2020-02-01	1

end_week	0	1	2020-04-25	2020-04-25	2020-04-25	1

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

covid_deaths	6	0.89	735.02	1801.11	0	54.50	153.00	519.00	10978.00	▇▁▁▁▁

total_deaths	0	1.00	13557.43	13996.83	856	3813.00	10721.00	17624.00	69341.00	▇▂▁▁▁

percent_expected_deaths	0	1.00	0.93	0.27	0	0.86	0.95	0.99	2.19	▁▂▇▁▁

pneumonia_deaths	0	1.00	1197.26	1453.17	41	277.00	769.00	1306.00	6076.00	▇▁▁▁▁

pneumonia_and_covid_deaths	10	0.81	355.81	759.51	0	30.50	65.00	296.00	4019.00	▇▁▁▁▁

all_influenza_deaths_j09_j11	3	0.94	116.58	142.24	14	30.50	87.50	125.50	850.00	▇▁▁▁▁

The U.S. Centers for Disease Control provides weekly summary and interpretation of key indicators that have been adapted to track the COVID-19 pandemic in the United States. Data is retrieved using the cdccovidview package from both COVIDView (https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html) and COVID-NET. Please see the indicated reference for all the caveats and precise meanings for each field. (https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html).

References

https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true

Author

Kieran Healy