CoronaNet Government Response Project data

Contains variables from the CoronaNet Government Response Project (https://coronanet-project.org), representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. See below for citation information.

coronanet

Format

A tibble with 18,074 rows and 33 variables:

record_id

character Unique identifier for each unique policy record

policy_id

double Identifier linking new policies with subsequent updates to policies

recorded_date

date-time When the record was entered into the dataset

date_announced

date When the policy was announced

date_start

date When the policy went into effect

date_end

date When the policy ended (if it has an explicit end date)

entry_type

character Whether the record is new, meaning no restriction had been in place before, or an update (restriction was in place but changed). Corrections are corrections to previous entries.

event_description

character A short description of the policy change

domestic_policy

double Indicates where policy targets an area within the initiating country (i.e. is domestic in nature)

type

character The category of the policy

type_sub_cat

character The sub-category of the policy (if one exists)

type_text

double Any additional information about the policy type (such as the number of ventilators/days of quarantine/etc.)

index_high_est

double The high (95% posterior density) estimate of the country policy activity score (0-100)

index_med_est

double The median (most likely) estimate of the country policy activity score (0-100)

index_low_est

double The low (95% posterior density) estimate of the country policy activity score (0-100)

index_country_rank

double The relative rank by each day for each country on the policy activity score

country

character The country initiating the policy

init_country_level

character Whether the policy came from the national level or a sub-national unit

province

character Name of sub-national unit

target_country

character Which foreign country a policy is targeted at (i.e. travel policies)

target_geog_level

character Whether the target of the policy is a country as a whole or a sub-national unit of that country

target_region

character The name of a regional grouping (like ASEAN) that is a target of the policy (if any)

target_province

character The name of a province targeted by the policy (if any)

target_city

character The name of a city targeted by the policy (if any)

target_other

logical Any geographical entity that does not fit into the targeted categories coded elsewhere

target_who_what

character Who the policy is targeted at

target_direction

character Whether a travel-related policy affects people coming in (Inbound) or leaving (Outbound)

travel_mechanism

character If a travel policy, what kind of transportation it affects

compliance

character Whether the policy is voluntary or mandatory

enforcer

character What unit in the country is responsible for enforcement

link

character A link to at least one source for the policy

iso3

character 3-digit ISO country code

iso2

character 2-digit ISO country code

Source

Cheng, Cindy, Joan Barcelo, Allison Hartnett, Robert Kubinec, and Luca Messerschmidt. 2020. “Coronanet: A Dyadic Dataset of Government Responses to the COVID-19 Pandemic.” BETAVersion 1.0. https://www.coronanet-project.org.

Details

Table: Data summary

Name	coronanet

Number of rows	18074

Number of columns	33

_______________________

Column type frequency:

character	23

Date	2

numeric	6

POSIXct	2

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

record_id	0	1.00	6	10	0	18074	0

policy_id	0	1.00	4	14	0	13547	0

entry_type	0	1.00	6	9	0	2	0

event_description	0	1.00	3	6463	0	13858	23

type	0	1.00	6	54	0	21	0

type_sub_cat	6756	0.63	5	140	0	108	0

country	0	1.00	4	32	0	198	0

init_country_level	0	1.00	8	31	0	5	0

province	10610	0.41	3	33	0	671	0

target_country	186	0.99	4	1905	0	558	0

target_geog_level	11965	0.34	13	62	0	6	0

target_region	17864	0.01	4	218	0	40	0

target_province	16672	0.08	3	1556	0	476	0

target_city	17588	0.03	2	286	0	296	0

target_other	18074	0.00	NA	NA	0	0	0

target_who_what	9539	0.47	5	61	0	13	0

target_direction	14757	0.18	7	16	0	3	0

travel_mechanism	12859	0.29	5	103	0	78	0

compliance	457	0.97	20	167	0	31	0

enforcer	0	1.00	6	219	0	178	0

link	146	0.99	1	1821	0	10052	0

iso3	48	1.00	1	3	0	197	0

iso2	69	1.00	1	2	0	196	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

date_start	0	1.00	2019-12-31	2024-05-29	2020-03-30	208

date_end	12926	0.28	2020-01-28	2024-05-29	2020-04-29	229

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

domestic_policy	0	1.00	0.85	0.36	0.00	1.00	1.00	1.00	1.00	▂▁▁▁▇

type_text	17942	0.01	19.97	11.46	2.00	14.00	20.00	28.00	85.00	▆▇▁▁▁

index_high_est	3872	0.79	53.89	2.36	48.29	52.47	53.58	54.98	70.74	▅▇▁▁▁

index_med_est	3872	0.79	52.23	1.93	45.42	51.24	52.31	53.46	57.59	▁▂▇▆▁

index_low_est	3872	0.79	50.59	2.52	37.19	49.72	51.11	52.13	56.26	▁▁▂▇▂

index_country_rank	3872	0.79	118.99	53.88	1.00	70.00	124.00	169.00	194.00	▂▅▃▅▇

Variable type: POSIXct

skim_variable	n_missing	complete_rate	min	max	median	n_unique

recorded_date	0	1	2020-03-28 05:38:17	2020-07-05 23:34:13	2020-04-26 08:58:53	13850

date_announced	14	1	2019-12-31 00:00:00	2020-07-05 00:00:00	2020-03-28 00:00:00	185

This file contains variables from the CoronaNet government response project, representing national and sub-national policy event data from more than 140 countries since January 1st, 2020. The data include source links, descriptions, targets (i.e. other countries), the type and level of enforcement, and a comprehensive set of policy types. For more detail on this data, you can see the codebook here. The format of the data is in country-day-record_id format. Some record_id values have letters appended to indicate that the general policy category type also has a value for type_sub_cat, which contains more detail about the policy, such as whether health resources refers to masks, ventilators, or hospitals. Some entries are marked as new_entry in the entry_type field for when a policy of that type was first implemented in the country. Later updates to those policies are marked as updates in entry_type. To see how policies are connected, look at the policy_id field for all policies from the first entry through updates for a given country/province/city. If an entry was corrected after initial data collection, it will read corrected in the entry_type field (the original incorrect data has already been replaced with the corrected data).

References

Robert Kubinec and Luiz Carvalho, "A Retrospective Bayesian Model for Measuring Covariate Effects on Observed COVID-19 Test and Case Counts", https://doi.org/10.31235/osf.io/jp4wk

Author

Kieran Healy