COVID-19 data for the USA, current as of Wednesday, July 22, 2020

A dataset containing US state-level data on COVID-19

covus

Format

A tibble with 131,665 rows and 7 columns

date

Date in YYYY-MM-DD format (date)

state

Two letter State abbreviation (character)

fips

State FIPS code (character)

data_quality_grade

character Data quality as assessed by COVID Tracking Project staff

measure

Outcome measure for this date

count

Count of measure

measure_label

character Outcome measure, suitable for use as a plot label

Source

The COVID-19 Tracking Project https://covidtracking.com

Details

Table: Data summary

Name	covus

Number of rows	131665

Number of columns	7

_______________________

Column type frequency:

character	5

Date	1

numeric	1

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1.00	2	2	0	56	0

fips	0	1.00	2	2	0	56	0

data_quality_grade	18717	0.86	1	2	0	6	0

measure	0	1.00	5	24	0	17	0

measure_label	0	1.00	6	54	0	17	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

date	0	1	2020-01-22	2020-07-21	2020-05-13	182

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

count	75229	0.43	67001.62	284794.9	0	152	1322	18005	6536932	▇▁▁▁▁

The measures tracked by the COVID tracking project are as follows:

measure	measure_label

positive	Positive Tests

negative	Negative Tests

pending	Pending Tests

hospitalized_currently	Currently Hospitalized

hospitalized_cumulative	Cumulative Hospitalized

in_icu_currently	Currently in ICU

in_icu_cumulative	Cumulative in ICU

on_ventilator_currently	Currently on Ventilator

on_ventilator_cumulative	Cumulative on Ventilator

recovered	Recovered

death	Deaths

total_tests_viral	Total number of PCR tests performed

positive_tests_viral	Total number of positive PCR tests

negative_tests_viral	Total number of negative PCR tests

positive_cases_viral	Total number of positive cases measured with PCR tests

death_confirmed	Deaths Confirmed

death_probable	Deaths Probable

Not all measures are reported by all states. The positive, negative, death, death_confirmed, and death_probable measures are cumulative counts. death_confirmed is the total number deaths of individuals with COVID-19 infection confirmed by a laboratory test. In states where the information is available, it tracks only those laboratory-confirmed deaths where COVID also contributed to the death according to the death certificate. death_probable is the total number of deaths where COVID was listed as a cause of death and there is not a laboratory test confirming COVID-19 infection.