Google Mobility Data

Data from Google's Community Mobility Reports on relative changes in movement trends by location type

google_mobility

Format

A data frame with 1,849,962 rows and 7 variables:

country_region_code

character Country Code

country_region

character Country or Region name

sub_region_1

character Subregion (e.g. US state) name

sub_region_2

character Subregion (e.g. US county) name

date

double Date in yyyy-mm-dd format

type

character Type of location. Values are retail, grocery (and pharmacy), parts, transit (hubs/stations), workplaces, and residential

pct_diff

integer Percent change from baseline activity

Source

Google LLC "Google COVID-19 Community Mobility Reports." https://www.google.com/covid19/mobility/ Accessed: 2020-05-09

Details

Location accuracy and the understanding of categorized places varies from region to region, so Google does not recommend using this data to compare changes between countries, or between regions with different characteristics (e.g. rural versus urban areas). Regions or categories are omitted if Google does not have have sufficient statistically significant levels of data for it. Changes for each day are compared to a baseline value for that day of the week. The baseline is the median value, for the corresponding day of the week, during the 5-week period Jan 3–Feb 6, 2020. What data is included in the calculation depends on user settings, connectivity, and whether it meets our privacy threshold. If the privacy threshold isn’t met (when somewhere isn’t busy enough to ensure anonymity) we don’t show a change for the day. As a result, you may encounter empty fields for certain places and dates. We calculate these insights based on data from users who have opted-in to Location History for their Google Account, so the data represents a sample of our users. As with all samples, this may or may not represent the exact behavior of a wider population.

References