Provisional COVID-19 Death Counts by Sex, Age, and State

Deaths involving coronavirus disease 2019 (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.

nchs_sas

Format

A tibble with 1,362 rows and 12 variables:

data_as_of

date Date of data release

start_week

date First week-ending date of data period

end_week

date Last week-ending date of data period

state

character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, and New York City, separate from New York state.

sex

character Sex

age_group

character Age group

covid_19_deaths

integer Deaths involving COVID-19 (ICD-code U07.1)

total_deaths

integer Deaths from all causes of death

pneumonia_deaths

integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)

pneumonia_and_covid_19_deaths

integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)

influenza_deaths

integer Influenza Deaths (ICD-10 codes J09-J11)

pneumonia_influenza_or_covid_19_deaths

integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)

Source

National Center for Health Statistics https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku

Details

Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis, and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have been suppressed in accordance with NHCS confidentiality standards.

References

https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku

Author

Kieran Healy