Provisional COVID-19 Death Counts by Sex, Age, and State

Deaths involving coronavirus disease 2019 (COVID-19), pneumonia, and influenza reported to NCHS by sex and age group and state.

nchs_sas

Format

A tibble with 1,362 rows and 12 variables:

data_as_of

date Date of data release

start_week

date First week-ending date of data period

end_week

date Last week-ending date of data period

state

character Jurisdiction of occurrence. One of: United States total, a US State, District of Columbia, and New York City, separate from New York state.

sex

character Sex

age_group

character Age group

covid_19_deaths

integer Deaths involving COVID-19 (ICD-code U07.1)

total_deaths

integer Deaths from all causes of death

pneumonia_deaths

integer Pneumonia Deaths (ICD-10 codes J12.0-J18.9)

pneumonia_and_covid_19_deaths

integer Deaths with Pneumonia and COVID-19 (ICD-10 codes J12.0-J18.9 and U07.1)

influenza_deaths

integer Influenza Deaths (ICD-10 codes J09-J11)

pneumonia_influenza_or_covid_19_deaths

integer Deaths with Pneumonia, Influenza, or COVID-19 (ICD-10 codes U07.1 or J09-J18.9)

Source

National Center for Health Statistics https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku

Details

Table: Data summary

Name	nchs_sas

Number of rows	1362

Number of columns	12

_______________________

Column type frequency:

character	3

Date	3

numeric	6

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1	4	20	0	54	0

sex	0	1	3	7	0	4	0

age_group	0	1	8	17	0	12	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

data_as_of	0	1	2020-07-15	2020-07-15	2020-07-15	1

start_week	0	1	2020-02-01	2020-02-01	2020-02-01	1

end_week	0	1	2020-07-11	2020-07-11	2020-07-11	1

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

covid_19_deaths	277	0.80	670.40	5052.61	0	0	18.0	139.00	121374	▇▁▁▁▁

total_deaths	130	0.90	6757.92	53323.21	0	70	429.0	2248.75	1387325	▇▁▁▁▁

pneumonia_deaths	330	0.76	762.77	5568.38	0	0	47.5	247.50	131332	▇▁▁▁▁

pneumonia_and_covid_19_deaths	294	0.78	293.62	2195.25	0	0	0.0	59.00	52431	▇▁▁▁▁

influenza_deaths	533	0.61	45.16	301.01	0	0	0.0	17.00	6519	▇▁▁▁▁

pneumonia_influenza_or_covid_19_deaths	359	0.74	1230.47	8858.99	0	12	76.0	385.00	205837	▇▁▁▁▁

Number of deaths reported in this table are the total number of deaths received and coded as of the date of analysis, and do not represent all deaths that occurred in that period. Data during this period are incomplete because of the lag in time between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. This delay can range from 1 week to 8 weeks or more. Missing values may indicate that a category has between 1 and 9 observed cases and have been suppressed in accordance with NHCS confidentiality standards.

References

https://data.cdc.gov/NCHS/Provisional-COVID-19-Death-Counts-by-Sex-Age-and-S/9bhg-hcku

Author

Kieran Healy