Provisional Death Counts for Coronavirus Disease (COVID-19): Weekly State-Specific Data Updates

This report provides a weekly summary of deaths with coronavirus disease 2019 (COVID-19) by select geographic and demographic variables. In this release, counts of deaths are provided by the race and Hispanic origin of the decedent.

nchs_wss

Format

A tibble with 11,501 rows and 12 variables:

data_as_of

date Date of analysis

start_date

date Start date of coverage

end_date

date End date of coverage

year

character Year. One of "2020", "2021", or "2020/2021".

month

dbl Month

obs_unit

character Unit of observation. One of: By Total, By Year, By Month.

state

character Geographical unit. One of: the United States, a U.S. State, the District of Columbia, or New York City. New York state measures do not include New York City

race_ethnicity

chr Race and ethnic group. One of: Non-Hispanic White, Non-Hispanic Black or African American, Non-Hispanic American Indian or Alaska Native, Non-Hispanic Asian, Non-Hispanic Native Hawaiian or Other Pacific Islander, Non Hispanic more than one race, Hispanic or Latino.

deaths

integer Count of deaths

dist_pct

double Distribution of COVID-19 deaths (%): Deaths for each group as a percent of the total number of COVID-19 deaths reported.

uw_dist_pop_pct

double Unweighted distribution of population (%): Population of each group as a percent of the total population.

wt_dist_pop_pct

double Weighted distribution of population (%): Population of each group as percent of the total population after accounting for how the race and Hispanic origin population is distributed in relation to the geographic areas impacted by COVID-19.

Source

National Center for Health Statistics https://data.cdc.gov/NCHS/Provisional-Death-Counts-for-Coronavirus-Disease-C/pj7m-y5uh

Details

Table: Data summary

Name	nchs_wss

Number of rows	11501

Number of columns	12

_______________________

Column type frequency:

character	6

Date	1

numeric	5

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

start_date	0	1	10	10	0	27	0

end_date	0	1	10	10	0	27	0

year	0	1	4	9	0	4	0

obs_unit	0	1	7	8	0	3	0

state	0	1	4	20	0	53	0

race_ethnicity	0	1	18	54	0	7	0

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

data_as_of	0	1	2022-03-16	2022-03-16	2022-03-16	1

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

month	1484	0.87	6.00	3.56	1	3.0	6.0	9.0	12.0	▇▃▃▃▅

deaths	3352	0.71	708.70	8957.20	0	0.0	19.0	135.0	622483.0	▇▁▁▁▁

dist_pct	3352	0.71	17.21	27.85	0	0.0	1.9	20.8	100.0	▇▁▁▁▁

uw_dist_pop_pct	0	1.00	14.28	23.57	0	0.9	3.1	12.7	92.7	▇▁▁▁▁

wt_dist_pop_pct	0	1.00	13.47	21.45	0	0.4	3.0	14.0	93.6	▇▁▁▁▁

The percent of deaths reported in this table are the total number of represent all deaths received and coded as of the date of analysis and do not represent all deaths that occurred in that period. Data are incomplete because of the lag in time between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. This delay can range from 1 week to 8 weeks or more, depending on the jurisdiction, age, and cause of death. Provisional counts reported here track approximately 1–2 weeks behind other published data sources on the number of COVID-19 deaths in the U.S. COVID-19 deaths are defined as having confirmed or presumed COVID-19, and are coded to ICD–10 code U07.1. Unweighted population percentages are based on the Single-Race Population Estimates from the U.S. Census Bureau, for the year 2018 (available from: https://wonder.cdc.gov/single-race-population.html). Weighted population percentages are computed by multiplying county-level population counts by the count of COVID deaths for each county, summing to the state-level, and then estimating the percent of the population within each racial and ethnic group. These weighted population distributions therefore more accurately reflect the geographic locations where COVID outbreaks are occurring. Jurisdictions are included in this table if more than 100 deaths were received and processed by NCHS as of the data of analysis.

Race and Hispanic-origin categories are based on the 1997 Office of Management and Budget (OMB) standards (1,2), allowing for the presentation of data by single race and Hispanic origin. These race and Hispanic-origin groups—non-Hispanic single-race white, non-Hispanic single-race black or African American, non-Hispanic single-race American Indian or Alaska Native (AIAN), non-Hispanic single-race Asian, and non-Hispanic single-race Native Hawaiian and Other Pacific Islander —differ from the bridged-race categories shown in most reports using mortality data.

New York State totals exclude New York City (provided in table separately).

Missing values may indicate that a category has between 1 and 9 observed cases and have been suppressed in accordance with NHCS confidentiality standards.

Author

Kieran Healy