Provisional Death Counts for Coronavirus Disease (COVID-19): Weekly State-Specific Data Updates

This report provides a weekly summary of deaths with coronavirus disease 2019 (COVID-19) by select geographic and demographic variables. In this release, counts of deaths are provided by the race and Hispanic origin of the decedent.

nchs_wss

Format

A tibble with 264 rows and 7 variables:

data_as_of

character Date of analysis

state

character Geographical unit. One of: the United States, a U.S. State, the District of Columbia, or New York City. New York state measures do not include New York City

group

character Population group

deaths

double

dist_pct

double COLUMN_DESCRIPTION

uw_dist_pop_pct

double COLUMN_DESCRIPTION

wt_dist_pop_pct

double COLUMN_DESCRIPTION

Source

National Center for Health Statistics https://data.cdc.gov/NCHS/Provisional-Death-Counts-for-Coronavirus-Disease-C/pj7m-y5uh

Details

Table: Data summary

Name	nchs_wss

Number of rows	264

Number of columns	7

_______________________

Column type frequency:

character	3

numeric	4

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

data_as_of	0	1	10	10	0	1	0

state	0	1	4	20	0	44	0

group	0	1	5	45	0	6	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

deaths	67	0.75	1228.47	5266.18	0.0	34.0	123.00	650.00	64073.0	▇▁▁▁▁

dist_pct	67	0.75	22.05	24.74	0.0	2.4	11.00	38.60	95.5	▇▂▂▁▁

uw_dist_pop_pct	0	1.00	16.67	24.14	0.1	1.8	4.75	16.88	93.1	▇▁▁▁▁

wt_dist_pop_pct	0	1.00	16.67	21.15	0.1	2.1	6.30	23.62	91.0	▇▁▁▁▁

The percent of deaths reported in this table are the total number of represent all deaths received and coded as of the date of analysis and do not represent all deaths that occurred in that period. Data are incomplete because of the lag in time between when the death occurred and when the death certificate is completed, submitted to NCHS and processed for reporting purposes. This delay can range from 1 week to 8 weeks or more, depending on the jurisdiction, age, and cause of death. Provisional counts reported here track approximately 1–2 weeks behind other published data sources on the number of COVID-19 deaths in the U.S. COVID-19 deaths are defined as having confirmed or presumed COVID-19, and are coded to ICD–10 code U07.1. Unweighted population percentages are based on the Single-Race Population Estimates from the U.S. Census Bureau, for the year 2018 (available from: https://wonder.cdc.gov/single-race-population.html). Weighted population percentages are computed by multiplying county-level population counts by the count of COVID deaths for each county, summing to the state-level, and then estimating the percent of the population within each racial and ethnic group. These weighted population distributions therefore more accurately reflect the geographic locations where COVID outbreaks are occurring. Jurisdictions are included in this table if more than 100 deaths were received and processed by NCHS as of the data of analysis.

Missing values may indicate that a category has between 1 and 9 observed cases and have been suppressed in accordance with NHCS confidentiality standards.

Author

Kieran Healy